Sort transform code into object-specific modules
I'm about to add more (and more complicated) implementations per object.
Having everything in one module was already a bit overwhelming, and this
would only have gotten worse.
