+++ crates/fj-kernel/src/algorithms/transform/surface.rs
@@ -1,0 +1,16 @@
+use fj_math::Transform;
+
+use crate::{objects::Surface, stores::Stores};
+
+use super::TransformObject;
+
+impl TransformObject for Surface {
+    type Transformed = Self;
+
+    fn transform(self, transform: &Transform, stores: &Stores) -> Self {
+        Self::new(
+            self.u().transform(transform, stores),
+            transform.transform_vector(&self.v()),
+        )
+    }
+}
